refactor: split providers/hooks and resolve all frontend lint errors
