+++ apps/web/src/components/PublicOnlyRoute.tsx
@@ -1,5 +1,5 @@
 import { Navigate } from "react-router-dom";
-import { useAuth } from "@/providers/AuthProvider";
+import { useAuth } from "@/providers/AuthContext";
 
 /**
  * Wraps a route that should only be accessible to unauthenticated users.
