+++ apps/web/src/pages/DashboardPage.tsx
@@ -1,4 +1,4 @@
-import { useAuth } from "@/providers/AuthProvider";
+import { useAuth } from "@/providers/AuthContext";
 import { Header } from "@/components/Header";
 import {
   Card,
