+++ apps/web/src/providers/ThemeContext.tsx
@@ -1,0 +1,21 @@
+import { createContext, useContext } from "react";
+
+export type Theme = "light" | "dark" | "system";
+
+export interface ThemeContextType {
+  theme: Theme;
+  setTheme: (theme: Theme) => void;
+}
+
+export const ThemeContext = createContext<ThemeContextType | null>(null);
+
+/**
+ * Hook to access and change the current theme.
+ */
+export function useTheme() {
+  const context = useContext(ThemeContext);
+  if (!context) {
+    throw new Error("useTheme must be used within a ThemeProvider");
+  }
+  return context;
+}
